requirements.txt
Flask
openai
python-dotenv
chromadb
pandas
werkzeug
boto3
